chore(engine): day 8 — stateless evaluatePolicy in @node9/policy-engine
The policy waterfall (DLP, smart rules, inline-exec / eval detection,
pipe-chain, SSH host extraction, provenance, sandbox, dangerous-words,
strict mode) now lives in packages/policy-engine/src/policy/. Pure
sub-helpers (pipe-chain, ssh-parser, flag-tables, analyzeShellCommand)
move with it.

Public API: stateless. Caller passes config, args, optional context
(agent, cwd, activeEnvironment) and an I/O hook bag (checkProvenance,
isTrustedHost). Engine stays free of fs/path/os/process imports — host
fs reads live in the proxy hooks.

Proxy src/policy/index.ts shrinks to a wrapper: getConfig() +
getActiveEnvironment() bind on entry, hooks are injected, then it calls
into the engine. shouldSnapshot and explainPolicy stay host-side
(Config-aware / heavy diagnostic I/O respectively). pipe-chain.ts,
ssh-parser.ts, flag-tables.ts become one-line re-exports so existing
test imports keep working.
